+++ src/engine/atoms/index.ts
@@ -37,6 +37,7 @@ import './添加延时锦囊';
 import './移除延时锦囊';
 import './拼点';
 import './无操作';
+import './展示事件';
 import './询问闪';
 import './询问杀';
 import './请求回应';